requirements.txt
Flask==2.3.2
anthropic==0.3.11
pymongo==3.12.0
tweepy==4.14.0
python-dotenv==1.0.0
Werkzeug==2.3.3
pydantic==1.10.12
requests==2.31.0